test/test_high_level_function.py
import unittest
import sys
import os
from google.protobuf.timestamp_pb2 import Timestamp
from datetime import datetime
from unittest.mock import MagicMock

sys.path.append(os.path.join(os.path.dirname(__file__), '..', 'client'))

import reddit_client
import reddit_pb2


sys.path.append(os.path.join(os.path.dirname(__file__), '..', 'high_level_function'))
import high_level_function


class TestRetrievePostAndMostUpvotedReply(unittest.TestCase):

    def test_retrieve_post_and_most_upvoted_reply(self):
        # Create mock responses for the functions used in the retrieve_post_and_most_upvoted_reply function
        create_post_response = MagicMock()
        create_post_response.post.post_id = "post123"

        create_comment_response = MagicMock()
        create_comment_response.comment.comment_id = "comment456"

        most_upvoted_comment_response = MagicMock()
        most_upvoted_comment_response.comment.comment_id = "comment789"
        most_upvoted_comment_response.replies = []

        most_upvoted_reply_response = MagicMock()
        most_upvoted_reply_response.comment_id = "reply123"
        most_upvoted_reply_response.score = 10

        # Create a mock API client
        api_client = MagicMock()

        # Mock the functions used in the retrieve_post_and_most_upvoted_reply function
        api_client.create_post.return_value = create_post_response
        api_client.vote_post.side_effect = [None, None, None]
        api_client.create_comment.return_value = create_comment_response
        api_client.vote_comment.side_effect = [None, None]

        # Mock the get_most_upvoted_comment_and_reply function
        api_client.list_top_comments.return_value = [most_upvoted_comment_response]
        most_upvoted_comment_response.replies = [most_upvoted_reply_response]

        # Call the function to be tested
        client = reddit_client.RedditClient()
        
        high_level_function.retrieve_post_and_most_upvoted_reply(client)

        # Add your assertions here to verify the expected behavior
        # For example:
        api_client.create_post.assert_called_once()
        api_client.vote_post.assert_called_with("post123", upvote=True)
        api_client.create_comment.assert_called_once()
        api_client.vote_comment.assert_called_with("comment456", upvote=True)

if __name__ == "__main__":
    unittest.main()
